fixed i386 get_stack_pos() to return size_t, fixed comment in QoreProgram.h
git-svn-id: svn+ssh://svn.code.sf.net/p/qore/code/qore/trunk@2071 1b0b35fe-f740-4c1b-b404-a338b32e513b

+++ include/qore/macros-i386.h
@@ -47,8 +47,8 @@ static inline void atomic_inc(volatile int *a) {
 #define HAVE_CHECK_STACK_POS
 #define STACK_DIRECTION_DOWN 1
 
-static inline long get_stack_pos() {
-   long addr;
+static inline size_t get_stack_pos() {
+   size_t addr;
    __asm("movl %%esp, %0" : "=g" (addr) );
    return addr;
 }
